requirements.txt
asgiref==3.8.1
black==24.4.2
botocore==1.31.85
cement==2.8.2
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
crispy-bootstrap5==2024.2
cryptography==42.0.8
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-debug-toolbar==4.4.5
django-environ==0.11.2
django-extensions==3.2.3
django-js-asset==2.2.0
django-multiselectfield==0.1.13
django-tinymce==4.1.0
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
filelock==3.15.4
flake8==7.1.0
future==0.16.0
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
pipdeptree==2.23.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pypiwin32==223
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-environ==0.4.54
pywin32==306
PyYAML==6.0.1
requests==2.32.3
semantic-version==2.8.5
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.3
watchdog==4.0.1
wcwidth==0.1.9
wheel==0.43.0
whitenoise==6.6.0
